reorganize css files and correct reducer names
