+++ src/components/App.js
@@ -1,7 +1,6 @@
 import React, { Component } from "react";
-import styles from "./App.css";
+import styles from "../componentStyles/App.css";
 import { BrowserRouter as Router, Route } from "react-router-dom";
-// import { Switch } from "react-router";
 import NavBar from "./NavBar"
 import NavBarMobile from "./NavBarMobile"
 import Chart from "./Chart"
@@ -11,7 +10,7 @@ class App extends Component {
   render() {
     return (
         <Router>
-          <div>
+          <div className={styles.fontAll}>
             <NavBar />
             <NavBarMobile />
             <Route exact path="/" component={Chart} />
