+++ src/components/Chart.js
@@ -1,6 +1,6 @@
 
 import React from 'react';
-import styles from "./Chart.css"
+import styles from "../componentStyles/Chart.css"
 
 export default function Stats(props) {
   return <div className={styles.chartContainer}>Chart div placeholder!</div>
