+++ src/components/Stats.js
@@ -1,6 +1,6 @@
 
 import React from 'react';
-import styles from "./Stats.css"
+import styles from "../componentStyles/Stats.css"
 
 export default function Stats(props) {
   return <div className={styles.statsContainer}>Stats div placeholder!</div>
