+++ src/store.js
@@ -1,9 +1,9 @@
 import {createStore, combineReducers} from "redux"
 import {reducer as formReducer} from "redux-form"
-import {drawerReducer} from "./reducer"
+import {choreReducer} from "./reducer"
 
 export default createStore(
   combineReducers({
     form: formReducer,
-    drawer: drawerReducer
+    chart: choreReducer
   }))
